# tweak modular decode speed vs binary size