Minor tweaks to improve the modular decode speed and binary size.

Before: (single-core, only decode speed is measured accurately)
```
station.png
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm  SSIMULACRA2   PSNR        pnorm       BPP*pnorm   QABPP   Bugs
----------------------------------------------------------------------------------------------------------------------------------------
jxl:d0:2          987  1487163   12.0458779   9.497  21.502         -nan 100.00000000  99.99   0.00000000  0.000000000000  12.046      0
jxl:d0:3          987  1416083   11.4701367   5.772   8.072         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.470      0
jxl:d0:4          987  1401819   11.3545996   1.489   6.190         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.355      0
jxl:d0:5          987  1390068   11.2594177   0.814   5.827         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.259      0
jxl:d0:6          987  1381828   11.1926744   0.537   5.269         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.193      0
jxl:d0:7          987  1370796   11.1033163   0.376   4.903         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.103      0
jxl:d0:8          987  1354026   10.9674809   0.156   4.872         -nan 100.00000000  99.99   0.00000000  0.000000000000  10.967      0
jxl:d0:9          987  1345638   10.8995389   0.031   4.594         -nan 100.00000000  99.99   0.00000000  0.000000000000  10.900      0
Aggregate:        987  1392819   11.2817001   0.711   6.604   0.00000000 100.00000000  99.99   0.00000000  0.000000000000  11.282      0
```

libjxl.so is 2726792 bytes (stripped release build, x64, clang 14)


After:
```
station.png
Encoding      kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm  SSIMULACRA2   PSNR        pnorm       BPP*pnorm   QABPP   Bugs
----------------------------------------------------------------------------------------------------------------------------------------
jxl:d0:2          987  1487163   12.0458779   9.769  22.046         -nan 100.00000000  99.99   0.00000000  0.000000000000  12.046      0
jxl:d0:3          987  1416083   11.4701367   5.820   8.228         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.470      0
jxl:d0:4          987  1401819   11.3545996   1.475   6.276         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.355      0
jxl:d0:5          987  1390068   11.2594177   0.799   5.890         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.259      0
jxl:d0:6          987  1381828   11.1926744   0.545   5.366         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.193      0
jxl:d0:7          987  1370796   11.1033163   0.384   5.023         -nan 100.00000000  99.99   0.00000000  0.000000000000  11.103      0
jxl:d0:8          987  1354026   10.9674809   0.153   4.935         -nan 100.00000000  99.99   0.00000000  0.000000000000  10.967      0
jxl:d0:9          987  1345638   10.8995389   0.032   4.692         -nan 100.00000000  99.99   0.00000000  0.000000000000  10.900      0
Aggregate:        987  1392819   11.2817001   0.715   6.725   0.00000000 100.00000000  99.99   0.00000000  0.000000000000  11.282      0
```

libjxl.so is 2710408 bytes


So it's a small speedup (1-2% faster modular decoding) and a small binary size reduction (-16kb).